pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "BookStore API"
authors = ["Emerson Penelli <emersonpenelli@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^3.2.8"
django-rest-framework = "^0.1.0"
django-extensions = "^3.1.3"
django-debug-toolbar = "^3.2.2"
psycopg2-binary = "^2.9.9"
gunicorn = "^20.1.0"
gitpython = "^3.1.43"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
